requirements.txt
pyyaml>=6.0
cryptography>=41.0.0
mnemonic>=0.20
bech32>=1.2.0
bip-utils>=2.9.0

